Another reorganization. Add raise_with_location
