+++ metafunctions/exceptions.py
@@ -1,3 +1,5 @@
 
+
 class MetaFunctionError(Exception):
     pass
+
